refactor(frontend): inline errors_type_filter into filters
